Refactor booking components into reusable tables and cards
